// src/app/Pages/Models/branchmaster.ts
export class Branchmaster {
    ID: number;

    NAME: string = '';

    SEQ_NO: number;
    SHORT_CODE: any;

    ADDRESS: any;

    COUNTRY_ID: any;
    COUNTRY_NAME: any;

    STATE_ID: any;
    STATE_NAME: any;

    DISTRICT_ID: any
    DISTRICT_NAME: any

    // CITY_ID: any;
    // CITY_NAME: any;

    PINCODE_ID: any;
    PINCODE: any;

    ORG_ID: any;
    CLIENT_ID: number;
    CODE: any;
    IS_ACTIVE: Boolean = true;
}